Keep a single SpannerUtils (common) to avoid name clashing
